requirements.txt
matplotlib~=3.8.0
numpy~=1.26.0
PyQt5~=5.15.9
minimalmodbus~=2.1.1
pyqtgraph~=0.13.3
pyserial~=3.5
pyModbusTCP~=0.2.0
XlsxWriter~=3.1.9
requests~=2.31.0
retry2
pymodbus~=3.5.4